requirements.txt
numpy
ollama
pyaudio
pyttsx3
SpeechRecognition[google-cloud]
PyAutoGUI
pynput
geocoder
simplegmail
google-cloud-speech
easyocr
keyboard
python-dotenv
flask